+++ lib/app/core/utils/utilidadesUtil.dart
@@ -19,6 +19,15 @@ import '../../presentation/widgets/custom_app_widgets.dart';
 
 class UtilidadesUtil {
 
+  static compartirPdf(String archivo) async {
+    try {
+      await Share.shareXFiles([XFile(archivo)], text: "PDF's");
+      //  await Share.shareFiles([archivo], text: 'Your PDF!');;
+    } catch (e) {
+      DialogosAwesome.getWarning(descripcion: "No se pudo cargar la página");
+    }
+  }
+
   static Future<Uint8List>  captureToImg(GlobalKey keyWidgetShared) async {
 
     Uint8List data=await Future.delayed(const Duration(milliseconds: 40), () async {