+++ lib/app_censo/data/datasources/remote/apis/host/url_api_provider_app_censo.dart
@@ -28,6 +28,8 @@ class UrlApiProviderAppCenso{
 
     UrlApiProviderApp _urlApiProviderApp = UrlApiProviderApp();
     String url = HostAppCenso.gethost();
+
+
     return _urlApiProviderApp.get(
       url: url,
       segmento: segmento,