+++ lib/app_censo/domain/request/request_censo.dart
@@ -7,3 +7,4 @@ part 'get_datos_procesos_activos.dart';
 part 'update_foto_perCenso_request.dart';
 part 'get_mesas_by_idUsuario.dart';
 part 'update_coordenadas_mesa_request.dart';
+part 'download_pdf_censo_request.dart';