+++ lib/app_censo/domain/usecases/censo_use_cases.dart
@@ -8,3 +8,4 @@ part 'save_census_person_photo_use_case.dart';
 part 'get_mesas_by_idUsuario_use_case.dart';
 part 'update_mesa_coordinates_use_case.dart';
 part 'fetch_data_history_census_use_case.dart';
+part 'download_pdf_censo_use_case.dart';